[IN-155] IN-175 feat: useOAuth hook
